Move reference helpers to models
