+++ src/parser.rs
@@ -362,85 +362,3 @@ fn empty_patch() -> Patch {
         engagement_id: None,
     }
 }
-
-impl Item {
-    /// All references associated with this item via its plugin ID.
-    pub fn references<'a>(&self, report: &'a NessusReport) -> Vec<&'a Reference> {
-        match self.plugin_id {
-            Some(pid) => report
-                .references
-                .iter()
-                .filter(|r| r.plugin_id == Some(pid))
-                .collect(),
-            None => Vec::new(),
-        }
-    }
-
-    /// Convenience helper returning CVE identifiers for this item.
-    pub fn cves<'a>(&self, report: &'a NessusReport) -> Vec<&'a str> {
-        self.references(report)
-            .into_iter()
-            .filter_map(|r| {
-                if r.source.as_deref() == Some("CVE") {
-                    r.reference.as_deref()
-                } else {
-                    None
-                }
-            })
-            .collect()
-    }
-
-    /// Convenience helper returning BID identifiers for this item.
-    pub fn bids<'a>(&self, report: &'a NessusReport) -> Vec<&'a str> {
-        self.references(report)
-            .into_iter()
-            .filter_map(|r| {
-                if r.source.as_deref() == Some("BID") {
-                    r.reference.as_deref()
-                } else {
-                    None
-                }
-            })
-            .collect()
-    }
-}
-
-impl Plugin {
-    /// All references for this plugin.
-    pub fn references<'a>(&self, report: &'a NessusReport) -> Vec<&'a Reference> {
-        match self.plugin_id {
-            Some(pid) => report
-                .references
-                .iter()
-                .filter(|r| r.plugin_id == Some(pid))
-                .collect(),
-            None => Vec::new(),
-        }
-    }
-
-    pub fn cves<'a>(&self, report: &'a NessusReport) -> Vec<&'a str> {
-        self.references(report)
-            .into_iter()
-            .filter_map(|r| {
-                if r.source.as_deref() == Some("CVE") {
-                    r.reference.as_deref()
-                } else {
-                    None
-                }
-            })
-            .collect()
-    }
-
-    pub fn bids<'a>(&self, report: &'a NessusReport) -> Vec<&'a str> {
-        self.references(report)
-            .into_iter()
-            .filter_map(|r| {
-                if r.source.as_deref() == Some("BID") {
-                    r.reference.as_deref()
-                } else {
-                    None
-                }
-            })
-            .collect()
-    }
-}
